chore: migrate from middleware to proxy convention
- Renamed src/middleware.ts to src/proxy.ts per Next.js 16 convention
- Updated error logging from 'middleware' to 'proxy'
- No functional changes, just following new naming convention
- Build verified successfully without deprecation warnings
