--- src/app/(private)/history/page.tsx
@@ -1,7 +1,0 @@
-export default function Page() {
-  return (
-    <div className="flex flex-col gap-4 md:gap-6">
-      <h1>History</h1>
-    </div>
-  )
-}
