+++ src/features/bills-to-pay-management/index.ts
@@ -1,3 +1,3 @@
-export * from "./model/use-bills-to-pay-management";
-export * from "./ui/bills-to-pay-list";
-export * from "./ui/payment-summary";
+export { BillsToPayManagement } from "./ui/bills-to-pay-management";
+export { BillsToPayList } from "./ui/bills-to-pay-list";
+export { PaymentSummary } from "./ui/payment-summary";
